chore: clean mypy strict — 0 errors (CL-g7e)
Closes the last 25 mypy errors. Now passes mypy strict cleanly.

Real fixes:
- src/risk/kill_switches.py + src/features/__init__.py: replaced
  `callable` (the builtin function) with proper `Callable[..., Any]`
  type hint. The lowercase `callable` is not a valid type — these
  were latent bugs that only showed up under strict.
- src/features/__init__.py _topo_sort: removed sentinel
  `FeatureSpec(name, [], None, 0)` that passed None for compute_fn
  (which is now correctly typed Callable, not None-able). Replaced
  with an explicit None-guard returning early.
- src/execution/broker.py: changed stream_prices ABC from
  `async def -> AsyncIterator` to `def -> AsyncIterator`. The async
  form gets read as `Coroutine[..., AsyncIterator]` by mypy, which
  doesn't match async generator implementations.
- src/security/vault_client.py: cast resp values to str / list[str]
  (json.loads returns Any; the API contract is str/list[str]).
- src/risk/stress_test.py: typed `results` as `dict[str, dict[str, float]]`
  and made literal 0s into 0.0 (was being inferred as int dict).
- Various explicit casts for ndarray returns and float coercions.

Bookkeeping:
- pyproject.toml: added _experimental/ to mypy exclude (auto-generated
  research output, strict typing here would block the research loop on
  every new candidate). Added statsmodels/sklearn/yfinance/wolfcrypt to
  ignore_missing_imports.

ruff src/ scripts/ also clean. 917 unit + 4 research integration tests
pass.

+++ src/strategies/rate_diff_mean_reversion.py
@@ -165,7 +165,7 @@ class RateDiffMRStrategy:
         if self._model is None:
             return None
         fair = self._model["alpha"] + self._model["beta"] * spread
-        return (price - fair) / self._model["residual_std"]
+        return float((price - fair) / self._model["residual_std"])
 
     def fit(self, train_data: pd.DataFrame) -> None:
         import statsmodels.api as sm
